+++ blueocean-rest-impl/src/main/java/io/jenkins/blueocean/service/embedded/OrganizationResolverImpl.java
@@ -1,12 +1,14 @@
 package io.jenkins.blueocean.service.embedded;
 
+import com.google.common.collect.ImmutableList;
 import hudson.Extension;
 import hudson.model.ItemGroup;
+import io.jenkins.blueocean.rest.factory.OrganizationResolver;
+import io.jenkins.blueocean.rest.model.BlueOrganization;
 import io.jenkins.blueocean.service.embedded.rest.OrganizationImpl;
 import jenkins.model.Jenkins;
 
 import java.util.Collection;
-import java.util.Collections;
 
 /**
  * Default implementation of {@link OrganizationResolver} for a master is to have everything in
@@ -30,7 +32,7 @@ public class OrganizationResolverImpl extends OrganizationResolver {
     }
 
     @Override
-    public OrganizationImpl get(String name) {
+    public BlueOrganization get(String name) {
         if (instance.getName().equals(name))
             return instance;
         else
@@ -38,8 +40,8 @@ public class OrganizationResolverImpl extends OrganizationResolver {
     }
 
     @Override
-    public Collection<OrganizationImpl> list() {
-        return Collections.singleton(instance);
+    public Collection<BlueOrganization> list() {
+        return ImmutableList.<BlueOrganization>of(instance);
     }
 
     @Override
